django/requirements.txt
asgiref==3.2.7
braintree==4.2.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.3.0
idna==2.10
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
